+++ allauth/socialaccount/providers/google/views.py
@@ -1,3 +1,5 @@
+import requests
+
 from django.conf import settings
 
 import jwt
@@ -11,6 +13,12 @@ from allauth.socialaccount.providers.oauth2.views import (
 
 from .provider import GoogleProvider
 
+
+IDENTITY_URL = (
+    getattr(settings, "SOCIALACCOUNT_PROVIDERS", {})
+    .get("google", {})
+    .get("IDENTITY_URL", "https://www.googleapis.com/oauth2/v2/userinfo")
+)
 
 ACCESS_TOKEN_URL = (
     getattr(settings, "SOCIALACCOUNT_PROVIDERS", {})
@@ -30,17 +38,40 @@ ID_TOKEN_ISSUER = (
     .get("ID_TOKEN_ISSUER", "https://accounts.google.com")
 )
 
+FETCH_USERINFO = (
+    getattr(settings, "SOCIALACCOUNT_PROVIDERS", {})
+    .get("google", {})
+    .get("FETCH_USERINFO", False)
+)
+
 
 class GoogleOAuth2Adapter(OAuth2Adapter):
     provider_id = GoogleProvider.id
     access_token_url = ACCESS_TOKEN_URL
     authorize_url = AUTHORIZE_URL
     id_token_issuer = ID_TOKEN_ISSUER
+    identity_url = IDENTITY_URL
+    fetch_userinfo = FETCH_USERINFO
 
     def complete_login(self, request, app, token, response, **kwargs):
+        data = None
+        id_token = response.get("id_token")
+        if id_token:
+            data = self._decode_id_token(app, id_token)
+            if self.fetch_userinfo and "picture" not in data:
+                info = self._fetch_user_info(token.token)
+                picture = info.get("picture")
+                if picture:
+                    data["picture"] = picture
+        else:
+            data = self._fetch_user_info(token.token)
+        login = self.get_provider().sociallogin_from_response(request, data)
+        return login
+
+    def _decode_id_token(self, app, id_token):
         try:
-            identity_data = jwt.decode(
-                response["id_token"],
+            data = jwt.decode(
+                id_token,
                 # Since the token was received by direct communication
                 # protected by TLS between this library and Google, we
                 # are allowed to skip checking the token signature
@@ -58,8 +89,16 @@ class GoogleOAuth2Adapter(OAuth2Adapter):
             )
         except jwt.PyJWTError as e:
             raise OAuth2Error("Invalid id_token") from e
-        login = self.get_provider().sociallogin_from_response(request, identity_data)
-        return login
+        return data
+
+    def _fetch_user_info(self, access_token):
+        resp = requests.get(
+            self.identity_url,
+            headers={"Authorization": "Bearer {}".format(access_token)},
+        )
+        if not resp.ok:
+            raise OAuth2Error("Request to user info failed")
+        return resp.json()
 
 
 oauth2_login = OAuth2LoginView.adapter_view(GoogleOAuth2Adapter)